+++ ggml.h
@@ -742,6 +742,13 @@ enum ggml_opt_result ggml_opt(
         struct ggml_tensor * f);
 
 //
+// quantization
+//
+
+size_t ggml_quantize_q4_0(float * src, void * dst, int n, int k, int qk, int64_t * hist);
+size_t ggml_quantize_q4_1(float * src, void * dst, int n, int k, int qk, int64_t * hist);
+
+//
 // system info
 //
